backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
beautifulsoup4==4.12.2
playwright==1.40.0
pandas==2.1.4
python-multipart==0.0.6
websockets==12.0
google-generativeai==0.3.2
python-dotenv==1.0.0
aiofiles==23.2.1